Add datetime import and query_results attribute for pricing comparison
This update adds `datetime` to the import list in `flight_monitor.py` to enable price comparison with the date. A new attribute, `query_results`, is added to the `FlightMonitor` class to store these comparisons. By appending a dictionary with pricing information and the date to `query_results`, it will allow for more robust comparison analyses. There were also minor grammar corrections in the docstring of `_process_incoming_flight_data`.