Add .env_example file for environment setup
A new file, named .env_example, has been added. This file consists of placeholder keys which need to be replaced with actual values for various services like Firebase and Telegram. These configurations are essential for the proper functioning of the application. This schema will aid in the setup of the local development environment avoiding issues with missing or wrongly named environment variables.